--- src/test/resources/cucumber.properties
@@ -1,1 +1,0 @@
-cucumber.publish.quiet=true
